BUGFIX When merging schemas, datasources from included file are ignored.
FUNCTION A override for field is possible - you can change any attribute for <field ../>, but be carefull: if you ommit any attribute in override definition, default value can be used, instead of previously defined value. 
Attributes "name" and "type" are required in overrided definition.

Example:
- main dbschema
<class name="Service">
    <table name="Service">
         ....
          <field name="ShortDescription" dbcolumn="short_description" type="String" size="8000" nullable="true" />
    </table>
</class>
- override:
<class name="Service">
    <table name="Service">
          <field name="ShortDescription" dbcolumn="short_description" type="String" size="4000" />
    </table>
</class>
result: field ShortDescription will have NULLABLE="FALSE" as this is a default value for this attribute.